django/requirements/base.txt
asgiref==3.2.10
attrs==20.1.0
Django==3.1
djangorestframework==3.11.1
importlib-metadata==1.7.0
iniconfig==1.0.1
mccabe==0.6.1
more-itertools==8.4.0
mysqlclient==2.0.1
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uuid==1.30
zipp==3.1.0
